feat: add modules task solution
